requirements.txt
Flask
Flask-cors
pyjwt
Flask-SQLAlchemy
cryptography
flask-bcrypt
python-dotenv
mysql-connector-python
pandas
sklearn
spacy>=3.0.0,<4.0.0
nltk